UI change: rather than rendering all menu items under one uiOutput() object, there are three dynamically created sidebar sections which are as follows. 1)  menuItems up to and including Community Profile tab; 2) the conditional community selection drop-down; 3) the menu items which are situated below the Community Profile menuItem
Sections 1 and 3 are rendered  in response to the user selecting an LA for this first time. Section 2 will hide/show depending on whether user is currently visiting the Community Profile tab.

Chnaged pickerInput to selectInput for the conditional selection drop-down.

Created an eventReactive object which stores a list of communities in the currently selected LA (named "communities_list")